+++ apps/client/src/app/(auth)/layout.tsx
@@ -21,7 +21,7 @@ export default function AuthLayout({ children }: { children: ReactNode }) {
         try {
           const response = await axios.get(endPoints.userDetails)
           const userId = response.data.user._id
-          const { email, privateKey, role, trialAvailable, selectedWorkspaceId } = response.data.user
+          const { email, privateKey, role, selectedWorkspaceId } = response.data.user
           const { name: selectedWorkspaceName, clientId, clientSecret } = response.data.workspace
           const hasActiveSubscription = response.data.hasActiveSubscription
 
@@ -36,7 +36,7 @@ export default function AuthLayout({ children }: { children: ReactNode }) {
 
           localStorage.setItem("clientId", clientId)
           localStorage.setItem("clientSecret", clientSecret)
-          dispatch("setUserState", { userId, email, privateKey, role, trialAvailable, selectedWorkspaceId, selectedWorkspaceName, clientId, clientSecret, hasActiveSubscription })
+          dispatch("setUserState", { userId, email, privateKey, role, selectedWorkspaceId, selectedWorkspaceName, clientId, clientSecret, hasActiveSubscription })
           setAuthorized(true)
         }